BI Bot/BI Bot/requirements.txt
numpy
pandas
streamlit
openai
langchain-openai
langchain-community
python-dotenv
chromadb
pyadomd
